+++ gsa/src/web/graphql/__mocks__/targets.js
@@ -15,27 +15,138 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-import {deepFreeze} from 'web/utils/testing';
+import {createGenericQueryMock, deepFreeze} from 'web/utils/testing';
 
-import {GET_TARGETS} from '../targets';
+import {
+  CLONE_TARGET,
+  DELETE_TARGETS_BY_IDS,
+  EXPORT_TARGETS_BY_IDS,
+  GET_TARGET,
+  GET_TARGETS,
+} from '../targets';
 
-const target1 = deepFreeze({
+export const mockTarget = deepFreeze({
   id: '159',
   name: 'target 1',
+  owner: 'admin',
+  comment: 'detailspage',
+  writable: true,
+  inUse: false,
+  creationTime: '2020-12-23T14:14:11+00:00',
+  modificationTime: '2021-01-04T11:54:12+00:00',
+  permissions: [{name: 'Everything'}],
+  hosts: ['123.234.345.456', ' 127.0.0.1'],
+  excludeHosts: ['192.168.0.1'],
+  maxHosts: 2,
+  portList: {
+    name: 'list',
+    id: 'pl1',
+  },
+  sshCredential: {
+    name: 'ssh',
+    id: 'ssh1',
+    port: 22,
+  },
+  smbCredential: {
+    name: null,
+    id: null,
+  },
+  esxiCredential: {
+    name: null,
+    id: null,
+  },
+  snmpCredential: {
+    name: null,
+    id: null,
+  },
+  tasks: [
+    {
+      name: 'task 1',
+      id: 't1',
+    },
+  ],
+  aliveTests: 'Schroedingers host',
+  allowSimultaneousIPs: true,
+  reverseLookupOnly: true,
+  reverseLookupUnify: false,
+  portRange: '1-5',
+  userTags: {
+    count: 1,
+    tags: [
+      {
+        id: '345',
+        name: 'target:unnamed',
+        value: null,
+        comment: null,
+      },
+    ],
+  },
 });
 
-const target2 = deepFreeze({
-  id: '199',
+export const inUseTarget = deepFreeze({
+  id: '346',
   name: 'target 2',
+  owner: 'admin',
+  comment: 'asdfg',
+  writable: true,
+  inUse: true,
+  creationTime: '2020-12-23T14:14:11+00:00',
+  modificationTime: '2021-01-04T11:54:12+00:00',
+  permissions: [{name: 'Everything'}],
+  hosts: ['127.0.0.1'],
+  excludeHosts: ['123.234.345.456', '192.168.0.1'],
+  maxHosts: 1,
+  portList: {
+    name: 'list',
+    id: 'pl2',
+  },
+  sshCredential: {
+    name: null,
+    id: null,
+  },
+  smbCredential: {
+    name: 'smb',
+    id: 'smb2',
+  },
+  esxiCredential: {
+    name: null,
+    id: null,
+  },
+  snmpCredential: {
+    name: null,
+    id: null,
+  },
+  aliveTests: 'Schroedingers host',
+  allowSimultaneousIPs: false,
+  reverseLookupOnly: true,
+  reverseLookupUnify: false,
+  portRange: '1-5',
+  tasks: [
+    {
+      name: 'task 1',
+      id: 't1',
+    },
+  ],
+  userTags: {
+    count: 1,
+    tags: [
+      {
+        id: '345',
+        name: 'target:unnamed',
+        value: null,
+        comment: null,
+      },
+    ],
+  },
 });
 
 const mockTargets = {
   edges: [
     {
-      node: target1,
+      node: mockTarget,
     },
     {
-      node: target2,
+      node: inUseTarget,
     },
   ],
   counts: {
@@ -53,6 +164,64 @@ const mockTargets = {
     lastPageCursor: 'target:3',
   },
 };
+
+export const noPermTarget = deepFreeze({
+  id: '159',
+  name: 'target 1',
+  owner: 'admin',
+  comment: 'detailspage',
+  writable: true,
+  inUse: false,
+  creationTime: '2020-12-23T14:14:11+00:00',
+  modificationTime: '2021-01-04T11:54:12+00:00',
+  permissions: [{name: 'get_targets'}],
+  hosts: ['123.234.345.456', ' 127.0.0.1'],
+  excludeHosts: ['192.168.0.1'],
+  maxHosts: 2,
+  portList: {
+    name: 'list',
+    id: 'pl1',
+  },
+  sshCredential: {
+    name: 'ssh',
+    id: 'ssh1',
+    port: 22,
+  },
+  smbCredential: {
+    name: null,
+    id: null,
+  },
+  esxiCredential: {
+    name: null,
+    id: null,
+  },
+  snmpCredential: {
+    name: null,
+    id: null,
+  },
+  tasks: [
+    {
+      name: 'task 1',
+      id: 't1',
+    },
+  ],
+  aliveTests: 'Schroedingers host',
+  allowSimultaneousIPs: true,
+  reverseLookupOnly: true,
+  reverseLookupUnify: false,
+  portRange: '1-5',
+  userTags: {
+    count: 1,
+    tags: [
+      {
+        id: '345',
+        name: 'target:unnamed',
+        value: null,
+        comment: null,
+      },
+    ],
+  },
+});
 
 export const createGetTargetsQueryMock = (variables = {}) => {
   const queryResult = {
@@ -72,3 +241,44 @@ export const createGetTargetsQueryMock = (variables = {}) => {
   };
   return [queryMock, resultFunc];
 };
+
+export const createGetTargetQueryMock = (
+  targetId = '159',
+  target = mockTarget,
+) => createGenericQueryMock(GET_TARGET, {target}, {id: targetId});
+
+const exportTargetsByIdsResult = {
+  exportTargetsByIds: {
+    exportedEntities: '<get_targets_response status="200" status_text="OK" />',
+  },
+};
+
+export const createExportTargetsByIdsQueryMock = (ids = ['159']) =>
+  createGenericQueryMock(EXPORT_TARGETS_BY_IDS, exportTargetsByIdsResult, {
+    ids,
+  });
+
+const bulkDeleteByIdsResult = {
+  deleteTargetsByIds: {
+    ok: true,
+  },
+};
+
+export const createDeleteTargetsByIdsQueryMock = (targetIds = ['159']) =>
+  createGenericQueryMock(DELETE_TARGETS_BY_IDS, bulkDeleteByIdsResult, {
+    ids: targetIds,
+  });
+
+export const createCloneTargetQueryMock = (
+  targetId = '159',
+  newTargetId = 'bar',
+) =>
+  createGenericQueryMock(
+    CLONE_TARGET,
+    {
+      cloneTarget: {
+        id: newTargetId,
+      },
+    },
+    {id: targetId},
+  );